Improve measured coverage and add coverage interpolation
This commit uses the C parser to construct coverage blocks based on the
source code. The coverage blocks generated from the binary only include
lines which are actually compiled and for which DWARF/BTF data was
generated. By merging the blocks from the the source code and the blocks
from the binary, we get a more accurate coverage percentage.

The second thing this commit implements is what I call "coverage
interpolation". This solves the issue of missing "optimized" lines.
Sometimes lines of code are optimized away, thus it won't appear as
a covered line in the report. However, if a line below it was executed,
we can infer that the line above and below it must also have been ran
(provided they they are part of the same code block without
possibilities of branching). This should result in code coverage that
is even more accurate.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>