Fixed null check return code to support socket programs
The instrumented code will set R0 to a return value and then exit the
program as a bounds check, which is required by the verifier. However,
the value we were using was `-1`, this is not an issue for most programs
but the verifier has logic to disallow undefined return codes from some
program types. So we changed the return value to `1` which seems to be
accepted everywhere.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>