improved performance when loading non-instrumented program
Loading the non-instrumented program to gather the verifier log happens
inside a for loop. We initially only use a 1MiB log size to limit the
memory usage and expand it using the loop if we detect we are hitting
the limit up to 256 MiB. However, we never exit this loop before we
hit the max memory amount which caused unnecessary delay and memory
usage. This commit fixes the issue.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>