Fixed handling of dead code
There were two situations involving dead code that coverbee did not
handle well. If the last part of a program contained dead code, the
list of merged states would be shorter than the list of instructions
since the verifier log never covered that part, causing a panic.

The second situation would be when there is dead code in the middle of
a program. In which case we would only have a "blank" state, and we
would start using registers which might already be in use.

If one of these situations is detected, we will still instrument the
program but assume we have no registers available which is the worst
case but also always save.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>